appium_tests/requirements.txt
Appium-Python-Client==4.3.0
pytest==8.3.5
pytest-html==4.1.1
pytest-ordering==0.6
selenium==4.27.1
